requirements.txt
colorama==0.4.6
nvidia-ml-py==13.595.45
# packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049096285/work
psutil==7.2.2
pynvml==13.0.1
PyYAML==6.0.3
tqdm==4.67.3
